refactor: move sharedialog to compornent directory
