+++ lib/ui/components/share_dialog.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:sound_classify_app/models/detail.dart';
+
+class ShareDetailDialog extends StatelessWidget {
+  const ShareDetailDialog({Key? key, required Detail this.detail})
+      : super(key: key);
+
+  final Detail detail;
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text('情報の共有'),
+      content: const Text('音声解析の結果と場所を共有しますか？'),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.of(context).pop(),
+          child: const Text('キャンセル'),
+        ),
+        TextButton(
+          onPressed: () => Navigator.of(context).pop(),
+          child: const Text('共有'),
+        ),
+      ],
+    );
+  }
+}
